dom/security/test/general/test_assert_about_page_no_csp.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Bug 1490977: Test Assertion if content privileged about: page has no CSP</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<iframe id="testframe"></iframe>
<script class="testbody" type="text/javascript">

  SimpleTest.waitForExplicitFinish();

  SimpleTest.expectAssertions(0, 1);
  // overruling the whitelist of about: pages that are allowed to load
  // without a CSP and making sure to hit the assertion within
  // AssertContentPrivilegedAboutPageHasCSP()

  SpecialPowers.pushPrefEnv({set: [
    ["csp.content_privileged_about_uris_without_csp", ""],
    ["csp.overrule_content_privileged_about_uris_without_csp_whitelist", true]]
  },
  function() {
    ok(true, "sanity: prefs flipped and test runs");
    
    let myFrame = document.getElementById("testframe");
    myFrame.src = "about:blank";
    // booom :-)
    SimpleTest.finish();
  });
</script>
</pre>
</body>
</html>
